requirements.txt
streamlit
azure-ai-documentintelligence
azure-core
pillow
opencv-python
pandas
numpy
plotly
python-dotenv